+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1+1
+
+- feat: add onTap for thumbnails
+- feat: current page highlighted
+
 # 0.1.0+1
 
 - feat: initial commit 🎉